String Level Loading Implementation
- Enemy_Spawner.cs is script attached to parent object of spawn points.
- Script controls all spawning of enemies, which state, which spawn
point, what formation, what type, etc.
- Utilities.cs added two methods, txtFileToStringArray that takes a
string path name and converts the txt file found at path to an array of
strings, separated by \n. The other method is getLevelTxtInstructions
which takes in just the level name and formats that path automatically
for passing as parameter to txtFileToStringArray.
- GameManager.cs now has a coroutine which takes in the argument of
each instruction and executes the code both spawning and delays in
waves. Coroutine is named SpawnWaveSet and the instruction string array
for that level is passed in.
- GameManager.cs also has a new function LoadLevelEnemies that takes in
the string of the level name, and calls the utilities.cs function to
read the level txt file and pass it into the coroutine.

- Most of the catches for bad exceptions should be in place. Left with
within the coroutine where the instruction itself has invalid
parameters.