Bug Fixes and Level Select implementation
- New key SELECTED_LEVEL is used to determine if there are selected
levels from the level select menu.
- If there are selected levels, this will override the GameManager’s
loading of the next level (highest cleared level + 1).
- Levels after the cleared selected level will be in sequence and will
not return to the highest cleared level until the user has beaten all
the way back up to that level, of the player exits to the main menu
which will delete the SELECTED_LEVEL data key.

- Fixed a bug whereby the going to the main menu from the pause menu
and going back to the game level scene after that will freeze the game.
This was due to timescale not being reseted back to 1.